refactor(tasks): Alter multichain deploy script to include ledger signer option
