+++ .env.example
@@ -2,6 +2,7 @@ RPC_URL_ETHEREUM_MAINNET="<RPC_URL>"
 ETHERSCAN_API_KEY="<ETHERSCAN_API_KEY>"
 SEQUENCER_ADDRESS="<SEQUENCER_ADDRESS>"
 ADMIN_SIGNER_PRIVATE_KEY="<SIGNER_PRIVATE_KEY>"
+LEDGER_ACCOUNT="<LEDGER_PUBLIC_KEY>"
 REPORTER_THRESHOLD="<REPORTER_THRESHOLD; default: 1>"
 ADMIN_THRESHOLD="<ADMIN_THRESHOLD; default: 1>"
 DEPLOY_WITH_SEQUENCER_MULTISIG_LOCAL="<true | false; default: true>"
