+++ libs/ts/contracts/package.json
@@ -34,14 +34,15 @@
     "@blocksense/sol-reflector": "workspace:*",
     "@nomicfoundation/hardhat-chai-matchers": "^2.0.8",
     "@nomicfoundation/hardhat-ethers": "^3.0.8",
+    "@nomicfoundation/hardhat-ledger": "^1.0.3",
     "@nomicfoundation/hardhat-network-helpers": "^1.0.12",
     "@nomicfoundation/hardhat-toolbox": "^5.0.0",
     "@nomicfoundation/hardhat-verify": "^2.0.12",
     "@nomiclabs/hardhat-ethers": "^2.2.3",
     "@openzeppelin/contracts": "^5.1.0",
-    "@safe-global/protocol-kit": "^4.0.2",
+    "@safe-global/protocol-kit": "^5.2.4",
     "@safe-global/safe-contracts": "1.4.1-2",
-    "@safe-global/safe-core-sdk-types": "^5.0.3",
+    "@safe-global/safe-core-sdk-types": "^5.1.0",
     "@safe-global/safe-deployments": "^1.37.22",
     "@typechain/ethers-v6": "^0.5.1",
     "@typechain/hardhat": "^9.1.0",
