+++ libs/ts/contracts/tasks/index.ts
@@ -1,3 +1,10 @@
+import './test-multichain-deploy';
+import './test-fund-ledger';
 import './multichain-deploy';
 import './verify';
-import './test-multichain-deploy';
+import './deployment-utils/deploy-multisig';
+import './deployment-utils/init-chain';
+import './deployment-utils/multisig-tx-exec';
+import './deployment-utils/deploy-contracts';
+import './deployment-utils/register-cl-adapters';
+import './deployment-utils/access-control';
