+++ libs/ts/contracts/tasks/types.ts
@@ -1,13 +1,13 @@
 import { EthereumAddress } from '@blocksense/base-utils/evm';
-import { JsonRpcProvider, Network, Wallet } from 'ethers';
+import { JsonRpcProvider, Network, Signer, Wallet } from 'ethers';
 
 export interface MultisigConfig {
-  signer: Wallet;
+  signer?: Wallet;
   owners: EthereumAddress[];
   threshold: number;
 }
 
-export interface NetworkConfig {
+interface NetworkConfigBase {
   rpc: string;
   provider: JsonRpcProvider;
   network: Network;
@@ -27,6 +27,22 @@ export interface NetworkConfig {
     safeWebAuthnSignerFactoryAddress: EthereumAddress;
   };
 }
+
+interface NetworkConfigWithLedger extends NetworkConfigBase {
+  ledgerAccount: Signer;
+  sequencerMultisig: Omit<MultisigConfig, 'signer'> & { signer?: undefined };
+  adminMultisig: Omit<MultisigConfig, 'signer'> & { signer?: undefined };
+}
+
+interface NetworkConfigWithoutLedger extends NetworkConfigBase {
+  ledgerAccount?: undefined;
+  sequencerMultisig: MultisigConfig;
+  adminMultisig: MultisigConfig;
+}
+
+export type NetworkConfig =
+  | NetworkConfigWithLedger
+  | NetworkConfigWithoutLedger;
 
 export enum ContractNames {
   SequencerMultisig = 'SequencerMultisig',
